+++ CHANGELOG.md
@@ -3,12 +3,13 @@
 ## [2.6.1] - 2018-04-22
 
 * Add front-page widget areas.
-* Add font awesome shortcodes (to do).
+* Add line awesome font shortcodes.
 * Add first, last, and widget-# CSS classes to widgets.
 * Add flexible widgets.
 * Add header search (to do).
 * Add Full Width Page Template for Beaver Builder and Elementor (to do).
 * Jetpack-like Related Posts without Jetpack (to do).
+* Add single post navigation.
 
 ## [2.6.0] - 2018-03-22
 * Fix product gallery bug in WooCommerce 3.0 update.