+++ CHANGELOG.md
@@ -2,12 +2,12 @@
 
 ## [2.6.1] - 2018-04-22
 
+* Force better (imho) layouts for WooCommerce.
 * Add front-page widget areas.
 * Add line awesome font shortcodes.
 * Add first, last, and widget-# CSS classes to widgets.
 * Add flexible widgets.
 * Add Full Width Page Template for Beaver Builder and Elementor (to do).
-* Jetpack-like Related Posts without Jetpack (to do).
 * Add single post navigation.
 
 ## [2.6.0] - 2018-03-22